store修正
