+++ frontend/pages/index.vue
@@ -2,7 +2,7 @@
   <v-container>
     <h1>Habi appで学習を習慣化しよう。</h1>
     <p>習慣化目標を記録するアプリです。</p>
-    <p>{{ }}</p>
+    <p>{{ $auth.user }}</p>
     <div>
       <HabitationsCard  v-for=" habitation in habitations" :key="habitation.id" :habitation="habitation" />
     </div>
