+++ frontend/plugins/persistedstate.js
@@ -1,0 +1,7 @@
+import createPersistedState from 'vuex-persistedstate'
+
+export default ({ store }) => {
+  createPersistedState({
+    key: 'HabiApp'
+  })(store)
+}
