--- frontend/store/modules/user.js
@@ -1,20 +1,0 @@
-export const strict = false
-
-export const state = () => ({
-  user: null
-})
-
-export const mutations = {
-  setUser(state, payload) {
-    state.user = payload
-  },
-  logout(state) {
-    state.user = null
-  }
-}
-
-export const actions = {
-  setUser(context, user) {
-    context.commit('setUser', user)
-  },
-}
